+++ src/components/beta/gux-button-slot/tests/gux-button-slot.e2e.ts
@@ -20,6 +20,10 @@ describe('gux-button-slot-beta', () => {
         html: '<gux-button-slot-beta accent="tertiary"><button type="button"accent="primary">Button</button></gux-button-slot-beta>'
       },
       {
+        description: 'should render ghost button',
+        html: '<gux-button-slot-beta accent="ghost"><button type="button"accent="primary">Button</button></gux-button-slot-beta>'
+      },
+      {
         description: 'should render invalid button',
         html: '<gux-button-slot-beta accent="invalid"><button type="button"accent="primary">Button</button></gux-button-slot-beta>'
       },
@@ -38,6 +42,10 @@ describe('gux-button-slot-beta', () => {
       {
         description: 'should render disabled tertiary button',
         html: '<gux-button-slot-beta accent="tertiary"><button type="button"accent="primary" disabled>Button</button></gux-button-slot-beta>'
+      },
+      {
+        description: 'should render disabled ghost button',
+        html: '<gux-button-slot-beta accent="ghost"><button type="button"accent="primary" disabled>Button</button></gux-button-slot-beta>'
       },
       {
         description: 'should render button with invalid accent',
@@ -60,6 +68,10 @@ describe('gux-button-slot-beta', () => {
         html: '<gux-button-slot-beta accent="tertiary"><input type="button"accent="primary" value="Button"/></gux-button-slot-beta>'
       },
       {
+        description: 'should render ghost input[type=button]',
+        html: '<gux-button-slot-beta accent="ghost"><input type="button"accent="primary" value="Button"/></gux-button-slot-beta>'
+      },
+      {
         description: 'should render invalid input[type=button]',
         html: '<gux-button-slot-beta accent="invalid"><input type="button"accent="primary" value="Button"/></gux-button-slot-beta>'
       },
@@ -78,6 +90,10 @@ describe('gux-button-slot-beta', () => {
       {
         description: 'should render disabled tertiary input[type=button]',
         html: '<gux-button-slot-beta accent="tertiary"><input type="button"accent="primary" disabled value="Button"/></gux-button-slot-beta>'
+      },
+      {
+        description: 'should render disabled ghost input[type=button]',
+        html: '<gux-button-slot-beta accent="ghost"><input type="button"accent="primary" disabled value="Button"/></gux-button-slot-beta>'
       },
       {
         description: 'should render input[type=button] with invalid accent',
@@ -100,6 +116,10 @@ describe('gux-button-slot-beta', () => {
         html: '<gux-button-slot-beta accent="tertiary"><input type="submit"accent="primary" value="Button"/></gux-button-slot-beta>'
       },
       {
+        description: 'should render ghost input[type=submit]',
+        html: '<gux-button-slot-beta accent="ghost"><input type="submit"accent="primary" value="Button"/></gux-button-slot-beta>'
+      },
+      {
         description: 'should render invalid input[type=submit]',
         html: '<gux-button-slot-beta accent="invalid"><input type="submit"accent="primary" value="Button"/></gux-button-slot-beta>'
       },
@@ -118,6 +138,10 @@ describe('gux-button-slot-beta', () => {
       {
         description: 'should render disabled tertiary input[type=submit]',
         html: '<gux-button-slot-beta accent="tertiary"><input type="submit"accent="primary" disabled value="Button"/></gux-button-slot-beta>'
+      },
+      {
+        description: 'should render disabled ghost input[type=submit]',
+        html: '<gux-button-slot-beta accent="ghost"><input type="submit"accent="primary" disabled value="Button"/></gux-button-slot-beta>'
       },
       {
         description: 'should render input[type=submit] with invalid accent',