+++ src/components/beta/gux-dropdown-v2/tests/gux-dropdown-v2.e2e.ts
@@ -64,7 +64,6 @@ describe('gux-dropdown-v2', () => {
       const focusEl = await page.find(':focus');
 
       expect(listbox.outerHTML).toContain(focusEl.outerHTML);
-      console.log(focusEl.outerHTML);
     });
 
     it('moves between options when arrow keys are pressed', async () => {