+++ src/components/legacy/gux-list-legacy/list-item/tests/gux-list-item.spec.ts
@@ -1,0 +1,20 @@
+import { newSpecPage } from '@stencil/core/testing';
+import { GuxListItemLegacy } from '../gux-list-item';
+
+describe('gux-list-item', () => {
+  let component: GuxListItemLegacy;
+
+  beforeEach(async () => {
+    const page = await newSpecPage({
+      components: [GuxListItemLegacy],
+      html: `<gux-list-item-legacy></gux-list-item-legacy>`,
+      language: 'en'
+    });
+
+    component = page.rootInstance;
+  });
+
+  it('should build', async () => {
+    expect(component).toBeInstanceOf(GuxListItemLegacy);
+  });
+});